core: Implement virtual memory mapping for cartridge ROM and RAM via MBCs
Implemented MBC1 virtual memory mapping for cartridge ROM and RAM. Based off a generic MBC trait which is also used to implement a fake non-mapping MBC if no MBC is present.

Mapping functions must be called for reads and writes and return either a physical address (in ROM or RAM) or an opaque register handle to a register inside the MBC. Methods are provided to read or write these internal registers using these handles.