core: Implement VRAM DMA Transfers
Implements VRAM DAM transfers, both general purpose and hblank. General purpose pauses the CPU for a number of ticks then executes the transfer. HBlank pauses the CPU to transfer a single block at the start of each HBlank.

Both are implemented with state machines that count down the number of ticks until the transfer or block is complete.

If double speed mode is ignored We can now succesfully boot Pokemon Crystal without any visual issues (though we crash when exiting the first room).