apu: Hook up audio playback to (stubbed) samples from APU
Set up scaffolding for audio playback. We now collect audio samples each frame and send them to the output device for playback at the end of the frame. The audio output is set up to be generic with a DefaultSystemAudioOutput used by default.

The rodio library is used for audio playback. We sample at 44.1 MHz which is a little more than every 95 ticks.